pyproject.toml
[project]
name = "simple-rag-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator==2.2.0",
    "pydantic==2.10.6",
    "python-dotenv==1.0.1",
    "streamlit==1.43.1",
]
